+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   - enclosure_group support for API300
   - firmware support for API300
   - storage_system support for API300
+  - storage_pool support for API300
 
 ### 1.1.0
   - Add support for client ENV variables